+++ src/components/icons/IconTwitch.tsx
@@ -1,0 +1,13 @@
+import { BaseIcon, type IconProps } from "./BaseIcon";
+
+// This SVG code is derived from FontAwesome (https://fontawesome.com/icons/twitch)
+export default function IconTwitch(props: IconProps) {
+  return (
+    <BaseIcon viewBox="0 0 512 512" {...props}>
+      <path
+        fill="currentColor"
+        d="M391.17,103.47H352.54v109.7h38.63ZM285,103H246.37V212.75H285ZM120.83,0,24.31,91.42V420.58H140.14V512l96.53-91.42h77.25L487.69,256V0ZM449.07,237.75l-77.22,73.12H294.61l-67.6,64v-64H140.14V36.58H449.07Z"
+      />
+    </BaseIcon>
+  );
+}
